chore: 死代码清理 —— scanner stub / chunkSlice 移测试包 / var _ 占位
纯清理，无功能影响。重审 (#132) 确认以下死代码/陈旧实现:

- internal/service/scanner.go: v1 scanner 删除后的空 stub (仅剩 license 头 +
  package 声明) → 删除文件
- internal/service/scanner_test.go: 测的是 domain.ValidateDeviceType (命名错位,
  在 package service 里测 package domain) → 移到 internal/domain/
  validate_device_type_test.go,与被测代码同包;保留有价值的测试逻辑
- routes.go chunkSlice: 仅被 batch_test.go 引用,无生产 caller → 移进
  batch_test.go 成为 test-only helper,从生产 routes.go 移除
- heartbeat_store.go / device_bridge.go / handler/scanner.go / totp.go:
  "var _ = slog.Default" / "var _ = errors.New" / "_ = user.ID" 等 import 保留器
  占位 —— 对应 import 都有真实引用 (slog/errors 在别处真用) 或值不用 (user.ID),
  占位纯属误导噪声 → 删除
- lldp_frame_real.go: "var _ = time.Second" + 注释 "kept for future variants"
  (YAGNI) → 删除占位 + 删除未用的 "time" import

验证: go build / go vet / gofmt / goimports 全干净; go test -race 全包通过
(routes / service / domain / config 等 20 个包全绿)。

Closes #132
