+++ internal/api/routes/batch_test.go
@@ -6,6 +6,24 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+// chunkSlice splits a slice into chunks of the given size. Test-only helper
+// (moved out of routes.go in #132 — it had no production callers, only this
+// test). Kept here so the batching logic stays under test.
+func chunkSlice[S any](items []S, batchSize int) [][]S {
+	if batchSize <= 0 {
+		return [][]S{items}
+	}
+	var chunks [][]S
+	for i := 0; i < len(items); i += batchSize {
+		end := i + batchSize
+		if end > len(items) {
+			end = len(items)
+		}
+		chunks = append(chunks, items[i:end])
+	}
+	return chunks
+}
+
 func TestChunkSlice(t *testing.T) {
 	// 100 items, batch size 50 → 2 chunks
 	items := make([]int, 100)
